requirements.txt
psycopg2-binary
pgvector
langchain
langchain-openai
langchain-core
langgraph
langchain-community
python-dotenv
